internal: add ranges to minicore
